HSEARCH-2841 Move the Elasticsearch IndexSearcher to a separate file
